Add Timer Code
and another small improvement I forgot. The timer will fire every three frames if done properly. To avoid reaching the integer limit while the timer goes infinitely, the value will reset every three seconds, or if it equals a value divisible by 180.

I couldn't figure out a timer, so I had to look through the Machinump's version of this to code it. Sorry!