Autosaving Works, and some other stuff I forgot
I'm back, baby! Real life got in the way for a bit, sorry about that. Anyways, simple: Autosaves every 15 seconds, and displays a message when doing so (please note that the current duration is not actually true, it just shows a warning. It loops to 1500 and then saves, but the message is displayed at 1000. So the message really isn't properly showing.